(maint) Fix plugin breaking change
Prior to this commit, the file_line logic was changed to only allow for
the new plugin defaults, which was across multiple lines. Unfortunately,
this causes any existing .zshrc to insert the plugin list in the wrong
place. This commit allows for the replacement of both the single line
plugin list and the multiline plugin list.